SuperEmbed + Selection Color added
FitVids has been replaced with SuperEmbed, jQuery has been removed.

Selection color has been added, there's 3 new variables in theme.yml for this. First one is nonetheme_selection_color which if sets to true will set the selection color to be the one from selection_color variable and not the highlight_color variable. There's also a variable for the color of the selecteed text.